+++ packages/js-marker-clusterer/directives/marker-cluster.ts
@@ -1,7 +1,7 @@
 import {Directive, Input, OnDestroy, OnChanges, OnInit, SimpleChange} from '@angular/core';
 
 import {ClusterManager} from '../services/managers/cluster-manager';
-import {MarkerManager, InfoWindowManager} from '@agm/core';
+import {MarkerManager, InfoWindowManager} from '@a7gm/core';
 
 import {ClusterOptions, ClusterStyle} from '../services/google-clusterer-types';