+++ packages/js-marker-clusterer/js-marker-clusterer.module.ts
@@ -1,5 +1,5 @@
 import {NgModule} from '@angular/core';
-import {AgmCoreModule} from '@agm/core';
+import {AgmCoreModule} from '@a7gm/core';
 import {AgmMarkerCluster} from './directives/marker-cluster';
 
 @NgModule({